Move parser logic into a separate file
